server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
python-socketio>=5.11.2
python-multipart>=0.0.9
requests>=2.32.3
google-genai>=1.0.0
langchain>=0.2.1
langchain-google-genai>=1.0.6
google-api-python-client>=2.133.0
google-auth>=2.30.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
aiohttp>=3.9.5
httpx>=0.27.0
